billing-app/app/requirements.txt
aio-pika
asyncpg
python-dotenv
fastapi
uvicorn